server-side caching implemented

+++ src/services/notes/repositories/note-repositories.js
@@ -1,11 +1,13 @@
 import { nanoid } from 'nanoid';
 import { Pool } from 'pg';
 import collaborationRepositories from '../../collaborations/repositories/collaboration-repositories.js';
+import CacheService from '../../../cache/redis-service.js';
 
 class NoteRepositories {
   constructor() {
     this.pool = new Pool();
     this.collaborationRepositories = collaborationRepositories;
+    this.cacheService = new CacheService();
   }
 
   async addNote({ title, body, tags, owner }) {
@@ -20,11 +22,24 @@ class NoteRepositories {
 
     const result = await this.pool.query(query);
 
+    await this.cacheService.delete(`notes:${owner}`);
+
     console.log(result.rows[0]);
     return result.rows[0];
   }
 
   async getNotes(owner) {
+    const cacheKey = `notes:${owner}`;
+
+    try {
+      const notes = await this.cacheService.get(cacheKey);
+      if (notes) {
+        return JSON.parse(notes);
+      }
+    } catch (error) {
+      console.error('Redis Error:', error);
+    }
+    // Cache miss, get from database
     const query = {
       text: `SELECT notes.* FROM notes
             LEFT JOIN collaborations ON collaborations.note_id = notes.id
@@ -32,7 +47,12 @@ class NoteRepositories {
             GROUP BY notes.id`,
       values: [owner],
     };
+
     const result = await this.pool.query(query);
+
+    // Save to cache
+    await this.cacheService.set(cacheKey, JSON.stringify(result.rows));
+
     return result.rows;
   }
 
@@ -51,22 +71,32 @@ class NoteRepositories {
     const updatedAt = new Date().toISOString();
 
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
     const result = await this.pool.query(query);
+
+    const owner = result.rows[0].owner;
+    if (result.rows[0]) {
+      await this.cacheService.delete(`notes:${owner}`);
+    }
 
     return result.rows[0];
   }
 
   async deleteNote(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
     const result = await this.pool.query(query);
+
+    const owner = result.rows[0].owner;
+    if (result.rows[0]) {
+      await this.cacheService.delete(`notes:${owner}`);
+    }
 
     return result.rows[0];
   }
